+++ Tools/Firmware/CMakeLists.txt
@@ -23,6 +23,8 @@ include_directories(
 
     ## Driver Includes ##
     ${ROOT_DIR}/Drivers/RcReceiver/Inc
+    ${ROOT_DIR}/Drivers/Common/Inc
+
 
     ## Boardfiles Includes ##
     ${ROOT_DIR}/Boardfiles/${FOLDER_NAME}/Drivers/${FAMILY_NAME}_HAL_Driver/Inc
@@ -60,6 +62,8 @@ set(ATTITUDE_MANAGER_CXX_SOURCES "${ATTITUDE_MANAGER}/Src/*.cpp"
 ## Drivers Sources ##
 set(DRIVERS_RcReceiver_DIR ${ROOT_DIR}/Drivers/RcReceiver/Src)
 set(DRIVERS_RcReceiver_CXX_SOURCES "${DRIVERS_RcReceiver_DIR}/*.cpp")
+set(DRIVERS_Common_DIR ${ROOT_DIR}/Drivers/Common/Src)
+set(DRIVERS_Common_CXX_SOURCES "${DRIVERS_Common_DIR}/*.cpp")
 
 ## Actually find the sources, NOTE: if you add a new source above, add it here! ##
 file(GLOB_RECURSE C_SOURCES ${HAL_DRIVERS_C_SOURCES}
@@ -69,7 +73,8 @@ file(GLOB_RECURSE C_SOURCES ${HAL_DRIVERS_C_SOURCES}
 message("C Sources: ${C_SOURCES}")
 file(GLOB_RECURSE CXX_SOURCES ${HAL_CORE_CXX_SOURCES}
                               ${ATTITUDE_MANAGER_CXX_SOURCES}
-                              ${DRIVERS_RcReceiver_CXX_SOURCES})
+                              ${DRIVERS_RcReceiver_CXX_SOURCES}
+                              ${DRIVERS_Common_CXX_SOURCES})
 message("C++ Sources: ${CXX_SOURCES}")
 
 ## Find the startup and linker script ##
